fix: Android build error - explicit cast for flog::debug parameters
- Cast duration.count() from long long to int for Android NDK compatibility
- Cast scanLogMinDurationMs to int to ensure consistent types
- Resolves C++ standard library conversion error in Android builds

+++ misc_modules/scanner/src/main.cpp
@@ -372,7 +372,7 @@ private:
                            currentTransmissionFreq / 1e6, durationSeconds);
             } else {
                 flog::debug("Scanner: Skipped short transmission {:.6f} MHz, duration: {}ms (< {}ms threshold)", 
-                           currentTransmissionFreq / 1e6, duration.count(), scanLogMinDurationMs);
+                           currentTransmissionFreq / 1e6, (int)duration.count(), (int)scanLogMinDurationMs);
             }
             isTrackingTransmission = false;
         }
